+++ src/main/java/br/com/neki/projeto/entity/Usuario.java
@@ -16,7 +16,7 @@ public class Usuario {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name="id")
-	private Long id;
+	private Integer id;
 	
 	@Column(name="login", nullable= false)
 	private String login;
@@ -32,7 +32,7 @@ public class Usuario {
 		super();
 	}
 
-	public Usuario(Long id, String login, String senha, LocalDate ultimoLoginData) {
+	public Usuario(Integer id, String login, String senha, LocalDate ultimoLoginData) {
 		super();
 		this.id = id;
 		this.login = login;
@@ -40,11 +40,11 @@ public class Usuario {
 		this.ultimoLoginData = ultimoLoginData;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
